pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "falk"
version = "0.1.0"
description = "Turn your data warehouse into an AI teammate with natural language queries powered by semantic layers"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "falk Contributors", email = "frederik.hagelund@gmail.com"}
]
keywords = [
    "data",
    "ai",
    "agent",
    "semantic-layer",
    "llm",
    "analytics",
    "duckdb",
    "slack-bot",
    "natural-language",
    "pydantic-ai",
    "metrics",
    "business-intelligence"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database",
    "Topic :: Office/Business",
]
dependencies = [
  "pydantic==2.12.5",
  "duckdb>=1.1.3",
  "PyYAML==6.0.2",
  "boring-semantic-layer>=0.3.7",
  # Pydantic AI agent + CLI chat + all main model providers
  "pydantic-ai-slim[openai,anthropic,google,groq,mistral,web,cli]==1.52.0",
  # Session state cache with TTL and eviction
  "cachetools>=5.0",
  # FastAPI wrapper and .env support
  "fastapi==0.128.0",
  "uvicorn[standard]>=0.38.0,<0.39.0",
  "python-dotenv==1.1.0",
  # Slack bot (socket mode)
  "slack-bolt>=1.18.0,<2.0.0",
  # Excel export
  "openpyxl>=3.1.0,<4.0.0",
  # Google Sheets export
  "google-api-python-client>=2.128.0,<3.0.0",
  "google-auth>=2.30.0,<3.0.0",
  # Chart generation (Plotly)
  "plotly>=6.1.1",
  "kaleido>=0.2.1,!=0.2.1.post1", # For PNG export
  # Pydantic Logfire observability (tracing, feedback)
  "logfire>=0.54.0",
  # Pydantic Evals — evaluation framework for agent testing
  "pydantic-evals>=0.1.0",
  # Postgres session store
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.1.0",
  # CLI framework
  "typer>=0.12.0",
  "rich>=14.0.0",
  # MCP (Model Context Protocol) server via FastMCP
  # Pinned to < 3 for stability (FastMCP 3.0 has breaking changes)
  "fastmcp>=0.5.0,<3",
]

[project.urls]
Homepage = "https://github.com/Fredehagelund92/Falk"
Documentation = "https://fredehagelund92.github.io/Falk/"
Repository = "https://github.com/Fredehagelund92/Falk"
Issues = "https://github.com/Fredehagelund92/Falk/issues"
Changelog = "https://github.com/Fredehagelund92/Falk/releases"

[project.optional-dependencies]
# Optional: Development tools
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.1.0",
  "pre-commit>=3.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/falk", "app"]

[tool.hatch.build.targets.wheel.shared-data]
"src/falk/scaffold" = "falk/scaffold"

[project.scripts]
falk = "falk.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/falk", "app"]
omit = ["*/tests/*", "*/scaffold/*"]

[tool.coverage.report]
fail_under = 55
show_missing = true

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/falk/prompt.py" = ["E501"]


